Refactor(woorim/boards): 컴포넌트 import 쉽게 되도록 최적화
