+++ app/src/components/index.js
@@ -1,7 +1,11 @@
-import Input from './Input';
-import Button from './Button';
-import Spinner from './Spinner';
-import Btn from './Btn'
-import Image from './Image'
+import Input from "./Input";
+import Button from "./Button";
+import Spinner from "./Spinner";
+import Btn from "./Btn";
+import Image from "./Image";
 
-export { Input, Button, Spinner, Btn, Image};
+// mains
+import MarketContainer from "./mains/market/MarketContainer";
+import LinearContainer from "./mains/linear/LinearContainer";
+
+export { Input, Button, Spinner, Btn, Image, MarketContainer, LinearContainer };
